Modified code for inheritance trees
Noticed old code was only working with one extending child type when the
item was generic. So I have changed the code to check the TypeSpec
signatures for GENERIC_INSTANCE types and then resolve those to see if
it is our class.

This works for the most part but there is still an issue where a single
entry is missing from the derived classes InheritedGenericClassTest4 -
this is a complicated type though so may have more to do with the way
the signature is produced.